`from fizzbuzz import ...` - change top level naming for python to `fizzbuzz`
