+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 
 members = [
-    "fizzbuzz",
+    "fizzbuzz-rust",
     "fizzbuzzo3"
 ]
