+++ fizzbuzz-rust/benches/bench_fizzbuzz.rs
