+++ fizzbuzz-rust/src/lib.rs
