+++ fizzbuzz-rust/tests/test_numbertypes.rs
