+++ fizzbuzz-rust/tests/test_vec.rs
