+++ fizzbuzzo3/Cargo.toml
@@ -10,6 +10,6 @@ crate-type = ["cdylib"]  # cdylib required for python import, rlib required for 
 
 [dependencies]
 pyo3 = { git = "https://github.com/MusicalNinjaDad/pyo3.git", branch = "pyo3-testing" }
-fizzbuzz = { path = "../fizzbuzz" }
+fizzbuzz = { path = "../fizzbuzz-rust" }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
